Функция перенесена в отдельный файл
