+++ functions.php
@@ -1,0 +1,12 @@
+<?php
+
+function cut_text (string $text, int $length = 300) {
+    if (mb_strlen($text) > $length) {
+        $text = mb_substr($text, 0, $length + 1);        
+        $end = mb_strlen(strrchr($text, ' '));
+        $text = mb_substr($text, 0, -$end) . '...';
+        }
+    return $text;
+}
+
+?>
